(extension) refactor popup route persistence to a route loader [DA-529]
Replace the two-phase (await hydrate before React mounts, then subscribe
after) implementation with React Router's native pattern:

- The index route's loader reads chrome.storage.session and returns
  redirect(persistedPath) or redirect('/mount'). Stale entries are
  cleared in the same loader. No window.location.hash mutation.
- RootRouter inlines the subscriber as a useEffect; no useMemo gymnastics
  on the router instance.

persistRoute.ts and its unit tests are gone; behaviour is covered by the
existing e2e specs in lifecycle/popupRoute.spec.ts. Net -207 lines.
