+++ packages/danmaku-anywhere/src/popup/router/RootRouter.tsx
@@ -1,11 +1,25 @@
-import { useEffect, useMemo } from 'react'
+import { useEffect } from 'react'
 import { RouterProvider } from 'react-router/dom'
-
-import { setupRoutePersistence } from './persistRoute'
-import { createPopupRouter } from './router'
+import { getStorageArea } from '@/common/storage/getStorageArea'
+import { POPUP_ROUTE_STORAGE_KEY, router } from './router'
 
 export const RootRouter = () => {
-  const router = useMemo(() => createPopupRouter(), [])
-  useEffect(() => setupRoutePersistence(router), [router])
+  useEffect(
+    () =>
+      router.subscribe((state) => {
+        if (state.navigation.state !== 'idle') {
+          return
+        }
+        const { pathname, search, hash } = state.location
+        const path = `${pathname}${search}${hash}`
+        if (path === '/') {
+          return
+        }
+        void getStorageArea('session').set({
+          [POPUP_ROUTE_STORAGE_KEY]: path,
+        })
+      }),
+    []
+  )
   return <RouterProvider router={router} />
 }
